Tried to solve a problem with HyperTester
The problem is the following: the language is lazy,
but the HyperTester is call-by-value, so sometimes
transformations may create something like this:

let x = E in Z

x is not used, but E will be computed by HyperTester.
And it may take a very long time, it may even diverge.

In this commit I've simpified the RunningContext class:
we don't recheck failed edged anymore. We don't throw
an exception when we detect an infinite loop,
we use a special value called ErrorBottom which is like _|_
but propagates even through constructors. There is also a Bottom
object which works like _|_