Now arity is not fixed and can be reduced
Arity of a node can be reduced before gluing to a
node with less arity or when an outgoing hyperedge
with less arity is being added. Arities of its parents
will be changed recursively.

Arity of a hyperedge can also be reduced. Strangely it turned
out to be very important to keep the number of nodes low.

Now I'm thinking of returning sets of used variables back, as
they are a generalization of arity.

Also added a test suite that just tests
evaluation of sample functions.