Implemented TerminationFilter
It filters out non-productive and non-coproductive
paths. Well, I'm not sure that it is perfectly correct
because it allows both guarded and structural recursion.
It may produce incorrect results if there is infinite data
somewhere. But I think in most cases it works. We
just need a good formal description of "most cases".