Now eq prover returns a proof tree; bug fixes
A proof tree can be converted to dot and it has
a method performGluing to perform actual gluing.

Also: the zeroth dest of a Let cannot be treated
as the rest of dests. May be we even should compute
a renaming for it first and then rearrange the
dests' tail.