Fixed some bugs, removed reduction to tick
"case S x of" used to be reduced by a transformation.
Now this kind of reduction is performed by the normalizer.
Also removed the ability of reduce these rhings to ticks.
I think ticks should be put into the code manually (or
automatically by a parser), not by some transformation.

Still have a problem of multiple defining hyperedges,
although I started fixing it.