Fixes; error -> unused
Error was a bad name for this type. Unused is better because
it represents code or value that can never be used, not code
that issues error. Now if we have an Id hyperedge like
f(x,y) = g(x,y,Unused)
then we know that for all x and y the third argument of g is
unused (it is represented as a variable -1 in the renaming).
And thus conversely g(x,y,z) = f(x,y)

Also moved some functions into the package object.

Also added a driving function. Driving is a complex transformation
that tries to create a defining hyperedge from the given node.

Also added some useful function to the parser, like parsing
a pair of expressions and assuming they are equal, and parsing
hyperedges.