Now dests should be invertible again
Well, turned out if we don't mention all
variables, the expression bacomes harder
to generalize.
Now a hyperedge should be interpreted as
a universally quantified statement:
  forall x y z . f(x) = g(x,y) + h(x,z)
Note that y and z are unused but they are still bound

(In the future we might need to specialize such
hyperedges)

The problem of transforming two hyperedges
when one of them has more variables in dests
than in the source is solved by preprocessing
hyperedge pairs (and adding those variables to
the dummy node's used set).

New transformations: letLetUp and caseGen
(just special cases of )

Trying to invent some new transformation manager.

And also fixed some bugs.