Added the ability of driving recommended nodes
When we perform merging by isomorphism we usually
get stuck at some unprovable equivalences. It may
be useful to drive the corresponding nodes to decide
whether they can really be equal.
Turned out that this is not useful at all because
it leads to explosion of the number of hyperedges.