Implemented new set of transformations (--drive2)
This new set of driving transformations doesn't replace
the scrutinee variable of a case-of with more complex
expressions, instead it works with let-caseof combinations.
In theory this may lead to more rewriting rules application
opportunities scince expressions are in a less reduced form.
Currently doesn't work very well, produces much more
hyperedges (useless, I suppose), and doesn't do much for
the rewriting rules examples. Maybe it was a wrong way.
The old set is used by default, the new one may be enabled
with the --drive2 flag.

Also enabled buffering by default. Doesn't affect performance, but
seems to work better with the new transformation set. Can be
disabled with --nobuf